python/agents/supply-chain/pyproject.toml
[project]
name = "supply-chain"
version = "0.1.0"
description = "power & energy supply chain AI agent"
authors = [
    {name = "Rohit Sroch",email = "rohitsroch@google.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "google-cloud-bigquery (>=3.36.0,<4.0.0)",
    "google-adk (>=1.0.0,<2.0.0)",
    "google-cloud-aiplatform[adk,agent-engines]>=1.132.0",
    "google-genai>=1.56.0",
    "db-dtypes (>=1.4.3,<2.0.0)",
    "pandas>=2.3.3",
    "matplotlib>=3.10.7,<4.0.0",
    "seaborn>=0.13.2,<0.14.0",
    "statsmodels>=0.14.5,<0.15.0",
    "absl-py>=2.3.1"
]

[project.optional-dependencies]
dev = [
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.132.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "google-adk[eval]>=1.5.0"
]

# The build system is updated to use hatchling, a standard build backend.
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
